requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
biopython==1.81
Brotli==1.0.9
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
comet_ml==3.52.1
configobj==5.0.9
contourpy==1.1.1
cycler==0.12.1
dulwich==0.22.4
einops==0.7.0
ema-pytorch==0.7.7
everett==3.1.0
fair-esm==2.0.0
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.2.0
huggingface-hub==0.36.0
Hydra==2.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lightning-utilities==0.11.9
llvmlite==0.41.1
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pip==24.0
pkgutil_resolve_name==1.3.10
prettytable==3.11.0
propcache==0.2.0
protobuf==5.29.5
psutil==5.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyDeprecate==0.3.2
Pygments==2.19.2
pyparsing==3.1.2
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.7.3
pytz==2025.2
PyYAML==6.0.2
rdkit==2023.3.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.2.1
scipy==1.6.2
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.37.0
setuptools==75.1.0
simplejson==3.20.1
six==1.16.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
torch_cluster==1.6.3+pt24cu121
torch_geometric==2.4.0
torch_scatter==2.1.2+pt24cu121
torch_sparse==0.6.18+pt24cu121
torchmetrics==0.11.4
tqdm==4.66.5
transformers==4.46.3
triton==3.0.0
typing_extensions==4.11.0
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.17.3
wurlitzer==3.1.1
xlsxwriter==3.2.9
yarl==1.15.2
zipp==3.20.2
